+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0
 
 - Add CORS handling.
+- `/stat` will return empty successful result instead of HTTP 404.
 
 
 # 0.1.0 (November 30, 2021)